requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
discord==2.3.2
discord.py==2.4.0
Django==5.1.3
djangorestframework==3.15.2
frozenlist==1.5.0
gunicorn==23.0.0
idna==3.10
multidict==6.1.0
packaging==24.2
propcache==0.2.0
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.2
toml==0.10.2
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
yarl==1.17.2